fix warnings

+++ levenshtein_distance.hpp
@@ -13,14 +13,14 @@ unsigned int levenshtein_distance(const T & s1, const T & s2)
     std::vector<int> col(len2+1);
     std::vector<int> prev_col(len2+1);
  
-    for (int i = 0; i < prev_col.size(); i++) {
+    for(size_t i=0; i<prev_col.size(); i++) {
         prev_col[i] = i;
     }
 
-    for (int i = 0; i < len1; i++) {
+    for(size_t i=0; i<len1; i++) {
         col[0] = i+1;
 
-        for (int j = 0; j < len2; j++){
+        for (size_t j=0; j<len2; j++) {
             col[j+1] = std::min(
                 std::min(prev_col[1 + j] + 1, col[j] + 1),
                 prev_col[j] + (s1[i] == s2[j] ? 0 : 1)
